tools: Remove nios2gen
All tools should be removed from the RTEMS source repository at some
point in time.  Tools with a BSD-style license will be moved to the
RTEMS tools repository.  Unfortunately, the this tool is RTEMS GPL
licensed.

If users of this tool still exist, they can reimplement the
functionality using a BSD-style license and add it to the RTEMS tools.

Close #3444.